Feature/exception with deleteall

+++ src/main/java/com/example/StudentApplication/service/ServiceHelper.java
@@ -1,14 +1,16 @@
 package com.example.StudentApplication.service;
 
 import com.example.StudentApplication.entities.Student;
-import com.example.StudentApplication.exception.CustomNotFoundException;
+import com.example.StudentApplication.exception.StudentNotFoundException;
+import com.example.StudentApplication.exception.StudentAlreadyExistsException;
 import com.example.StudentApplication.models.APIResponse;
 import com.example.StudentApplication.models.StatusEnum;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
-import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,17 +24,24 @@ public class ServiceHelper {
     }
 
     public APIResponse<List<Student>> findall() {
-
         APIResponse<List<Student>> apiResponse = new APIResponse<>();
-        apiResponse.setStatus(StatusEnum.SUCCESS);
-        apiResponse.setMessage("List of student successfully fetched");
-        apiResponse.setResponse(studentService.findAll());
-        return apiResponse;
+        Optional<List<Student>> findAllOptional = studentService.findAll();
+        if (findAllOptional.isPresent()) {
+            List<Student> liststud = findAllOptional.get();
+            if (!liststud.isEmpty()) {
+                apiResponse.setMessage("List of students successfully fetched");
+                apiResponse.setResponse(findAllOptional.get());
+                apiResponse.setStatus(StatusEnum.SUCCESS);
+            } else {
+                throw new StudentNotFoundException("List not found, kindly feed in values");
+            }
 
+        }
+        return apiResponse;
     }
 
     public APIResponse<Student> findById(int stud_id) {
-        //TODO: check if the repo response is null, if so throw custom not found exception
+
         Optional<Student> studentIdOptional = studentService.findById(stud_id);
         APIResponse<Student> apiResponse = new APIResponse<>();
         if (studentIdOptional.isPresent()) {
@@ -40,39 +49,77 @@ public class ServiceHelper {
             apiResponse.setResponse(studentIdOptional.get());
             apiResponse.setStatus(StatusEnum.SUCCESS);
         } else {
-           /* apiResponse.setStatus(StatusEnum.FAILURE);
-            apiResponse.setMessage("Student with id - " +stud_id +" details can't be found:");*/
-            throw new CustomNotFoundException("Student with id - " + stud_id + " details can't be found:");
-            //TODO: Throw a exception
+            throw new StudentNotFoundException("Student with id - " + stud_id + " details can't be found:");
         }
         return apiResponse;
-}
+    }
 
-    public APIResponse<Student> addStudent(Student student){
+    public APIResponse<Student> addStudent(Student student) {
 
+        Optional<Student> existingStudentNameOptional = studentService.findByStud_name(student.getName());
         APIResponse<Student> apiResponse = new APIResponse<>();
-        apiResponse.setResponse(studentService.addStudent(student));
-        apiResponse.setMessage("Inserted Successfully");
-        apiResponse.setStatus(StatusEnum.SUCCESS);
+        if (existingStudentNameOptional.isEmpty()) {
+            apiResponse.setResponse(studentService.addStudent(student));
+            apiResponse.setMessage("Inserted Successfully");
+            apiResponse.setStatus(StatusEnum.SUCCESS);
+        } else {
+            throw new StudentAlreadyExistsException("Student with name: " + student.getName() + " already exist");
+        }
+        return apiResponse;
+
+
+    }
+
+    public APIResponse<Student> deleteStudent(int stud_id) {
+        Optional<Student> existingDeletedStudentOptional = studentService.findById(stud_id);
+        APIResponse<Student> apiResponse = new APIResponse<>();
+        if (existingDeletedStudentOptional.isPresent()) {
+            apiResponse.setResponse(studentService.deleteStudent(stud_id));
+            apiResponse.setMessage("Student Deleted Successfully");
+            apiResponse.setStatus(StatusEnum.SUCCESS);
+        } else {
+            throw new StudentNotFoundException("Student with id: " + stud_id + " not found");
+        }
         return apiResponse;
     }
 
-    public APIResponse<Student> deleteStudent(int stud_id){
+   /* public APIResponse<Student> deleteAllStudents() {
 
+        Optional<List<Student>> studentOptional = studentService.findAll();
         APIResponse<Student> apiResponse = new APIResponse<>();
-        apiResponse.setResponse(studentService.deleteStudent(stud_id));
-        apiResponse.setMessage("Student Deleted Successfully");
-        apiResponse.setStatus(StatusEnum.SUCCESS);
+        if (studentOptional.isPresent()) {
+            List<Student> listStudents = studentOptional.get();
+            if (listStudents.isEmpty()) {
+                throw new StudentNotFoundException("Student table is empty");
+            } else {
+                studentService.deleteAllStudents();
+                apiResponse.setMessage("Students Deleted Successfully");
+                apiResponse.setStatus(StatusEnum.SUCCESS);
+            }
+        }
+        return apiResponse;
+    }*/
+    //directly delete everything, if any exception thrown,
+    // catch(if half record found messge) and handle. if no exception, all record deleted successfully message.
+
+    public APIResponse<Student> deleteAllStudents() {
+
+        Optional<List<Student>> studentOptional = studentService.findAll();
+        APIResponse<Student> apiResponse = new APIResponse<>();
+        if (studentOptional.isPresent()) {
+            List<Student> listStudents = studentOptional.get();
+            if (listStudents.isEmpty()) {
+                throw new StudentNotFoundException("Student table is empty");
+            } else {
+                studentService.deleteAllStudents();
+                apiResponse.setMessage("Students Deleted Successfully");
+                apiResponse.setStatus(StatusEnum.SUCCESS);
+            }
+        }
         return apiResponse;
     }
 
-    public APIResponse<Student> deleteAllStudents(){
-
-        APIResponse<Student> apiResponse = new APIResponse<>();
-        apiResponse.setResponse(studentService.deleteAllStudents());
-        apiResponse.setMessage("Students Deleted Successfully");
-        apiResponse.setStatus(StatusEnum.SUCCESS);
-        return apiResponse;
-
+    public long countStudents() {
+        return studentService.countStudents();
     }
 }
